# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/push/deposit": {
            "post": {
                "description": "Пополняет баланс указанного аккаунта",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Пополнение счета",
                "parameters": [
                    {
                        "description": "Данные пополнения",
                        "name": "deposit",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.DepositRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Пополнение успешно",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Некорректные данные запроса",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Ошибка пополнения счета",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/push/transfer": {
            "post": {
                "description": "Осуществляет перевод средств между счетами",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transactions"
                ],
                "summary": "Перевод средств",
                "parameters": [
                    {
                        "description": "Данные для перевода",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.TransferRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Push-уведомление отправлено",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Некорректные данные запроса",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Ошибка перевода",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/push/withdraw": {
            "post": {
                "description": "Позволяет пользователю снять средства со счета",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accounts"
                ],
                "summary": "Снятие средств",
                "parameters": [
                    {
                        "description": "Данные для снятия средств",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.WithdrawRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Снятие успешно",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Некорректные данные запроса",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Ошибка снятия средств",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.DepositRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "main.TransferRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "from_account": {
                    "type": "string"
                },
                "to_account": {
                    "type": "string"
                }
            }
        },
        "main.WithdrawRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "user_id": {
                    "type": "string"
                }
            }
        }
    }
}